Fix the cppcoreguidelines-* and readability-* findings from clang-tidy
- TrainingConfig's and SamplingOptions' hyperparameters, and the
  struct stat / task_basic_info out-parameters, are value-initialized
  instead of left indeterminate until assigned
  (cppcoreguidelines-pro-type-member-init).
- ThreadPool, the dashboard's RawTerm and token_file's Fd release a
  resource in their destructor but could be copied; their copy
  operations are deleted, and Fd gets a constructor. Command and
  MappedTokenDataset, which already deleted copying, spell out their
  defaulted destructor (cppcoreguidelines-special-member-functions).
- Command::parse and BPETokenizer::decode name their parameter the same
  in declaration and definition
  (readability-inconsistent-declaration-parameter-name).
- A redundant cast of std::atoll's result and a redundant "" string
  initializer go. The SA_RESETHAND cast stays with a targeted NOLINT:
  it is a no-op on macOS but needed on glibc, where the macro is
  unsigned (readability-redundant-casting,
  readability-redundant-string-init).

Determinism probe, train-fast metrics, greedy decoding and val loss
1.6929 are bit-identical.

+++ include/grad/transformer/blas_wrapper.h
@@ -28,7 +28,7 @@ inline bool metal_worthwhile(size_t flops) {
     static const long long threshold = [] {
         if (const char* value =
                 env::lookup("GRAD_METAL_THRESHOLD", "TRANSFORMER_METAL_THRESHOLD")) {
-            return static_cast<long long>(std::atoll(value));
+            return std::atoll(value);
         }
         return 10LL * 1000 * 1000 * 1000;
     }();
